requirement.txt
flask
spacy
scikit-learn
pdfplumber
python-docx
openai
nltk